Combined Publish, Retrieve, and Authentication into Services. Created tests for new files
